requirements.txt
python-dotenv
openai
aiogram==3.17.0
asyncio